Fix id field issue caused by numerify
Since the application was having trouble using `randomNumber()` for ints longer than a certain length, I decided to fudge it using `numerify()` and then casting to an int. This worked for the most part. But if a string is generated with leading 0s, they get removed when cast to int.

To get around this, `regexify()` is now used instead to force the first digit to be anything but 0.