Use brace expressions to pass data
The docs say to prefix dynamic attributes with : to treat the content as an expression, but PhpStorm was looking at that like Vue syntax and having problems with it.

To work around this, I removed the : prefix and switched back to Blade's normal braces.

Relevant doc: https://laravel.com/docs/9.x/blade\#passing-data-to-components

Next time I'll see if the issue is my Vue extension.