Revert to storing short path in DB
Read an important comment on the dev.to article I was following: https://dev.to/hosamalzagh/comment/1c0p4

So to delete the image file, the short path is needed to that file.

This also means it's good to save which disk the file is stored on. This can be used to dynamically generate the image URL based on both pieces of info.