issue 88: image lightbox modal (#115)
* Start on #88

* Found a way to restrict image height

This requires setting a hard max height, even if calculated. Next: can the --spacing be updated without rerendering the full component? And can it be calculated from the caption height and lightbox padding?

* Calculate element gap and image max height

* Caption will just overlay

This is the simplest path forward and bypasses needing extra calculations

* Upgrade deps in hope of fixing weird bug

Vite's Vue plugin kept complaining about some weird error that with 'sys', but it wasn't clear what the error was.

All that was known was that Vue was having trouble with using imported types in defineProps.

* Create captioned image component

* Create open, close, and update functions

* Make lightbox work on post view pages

* Fix width

* Fix transition

* Manually build data for lightbox

* wrap img and figcaption in figure

This is the correct way to use the img and figcaption together.

* Get img and figcaption data from figure

* Add some border and outline rounding

* Round images in admin image and project views

This is only for consistency.

* Auto-wrap/caption markdown images