Begin fixing GA init problems
Before I was loading the old GA component—which is fine for Blade, but it doesn't work for Livewire due to slots replacing stacks.

So GA will now be configured in a Volt component as well.

I think some of the leftover vanilla GA stuff will need to become Alpine components.